🔨 Refactor import paths for grapher_model
