+++ apps/backport/backport.py
@@ -17,10 +17,10 @@ from apps.backport.datasync.data_metadata import (
 )
 from apps.backport.datasync.datasync import upload_gzip_dict
 from etl import config, paths
-from etl import grapher_model as gm
 from etl.backport_helpers import GrapherConfig
 from etl.db import get_engine, read_sql
 from etl.files import checksum_str
+from etl.grapher import model as gm
 from etl.grapher.io import variable_data_df_from_s3
 from etl.snapshot import Snapshot, SnapshotMeta
