+++ apps/wizard/app_pages/chart_diff/chart_diff.py
@@ -12,10 +12,10 @@ from structlog import get_logger
 
 from apps.wizard.utils import get_staging_creation_time
 from apps.wizard.utils.io import get_all_changed_catalog_paths
-from etl import grapher_model as gm
 from etl.config import OWID_ENV
 from etl.db import read_sql
 from etl.git_helpers import get_changed_files, log_time
+from etl.grapher import model as gm
 
 log = get_logger()
