+++ etl/grapher/to_db.py
@@ -36,7 +36,7 @@ from etl import config
 from etl.db import get_engine, production_or_master_engine
 from etl.grapher import helpers as gh
 
-from .. import grapher_model as gm
+from . import model as gm
 
 log = structlog.get_logger()
